// src/app/modulos/atendimento/components/registrar/model/registrar-atendimento.ts
export interface RegistrarAtendimento {
  anoDocumento?: string;
  beneficiado?: string;
  codigo?: number;
  contato?: number;
  data?: any;
  dataFimAtendimento?: string;
  dataInicioAtendimento?: string;
  dataPrazo?: string;
  descricao?: string;
  digitoRap?: number;
  id?: number;
  idAssunto?: number;
  idAtuacao?: number;
  idBairro?: number;
  idBairroContato?: number;
  idCadastrado?: number;
  idCadastradoPor?: number;
  idCategoria?: number;
  idFinalizado?: number;
  idOcorrencia?: number;
  idOrgao?: number;
  idOrigem?: number;
  idPrazo?: number;
  idRegiao?: number;
  idResponsavelAtendimento?: number;
  idResponsavelOcorrencia?: number;
  idRetornar?: number;
  idTipoOcorrencia?: number;
  logradouro?: string;
  numeroAtendimento?: number;
  numeroDocumento?: number;
  numeroRap?: number;
  numeroSei?: number;
  numeroproc?: number;
  observacao?: number;
  opcaoNaoListar?: string;
  solicitante?: string;
  solicitantes?: string;
  status?: string;
}